optimize general styling and introduce react marque for horizontal slider
