+++ src/app/speakers/components/Hero.tsx
@@ -3,6 +3,7 @@ import Image from 'next/image';
 import styles from '../styles/Hero.module.scss';
 import { speakersSlideData } from '../mockData/mock';
 import { useEffect, useState } from 'react';
+import Marquee from 'react-fast-marquee';
 
 const Hero = () => {
   const [screenWidth, setScreenWidth] = useState<number | null>(null);
@@ -55,7 +56,11 @@ const Hero = () => {
       </div>
       {/* DISPLAY VERTICAL OR HORIZONTAL SLIDERS BASED ON SCREEN WIDTH */}
       {screenWidth > 768 ? (
-        <div className={styles.verticalMarquee}>
+        // Vertical React fast Marquee is buggy. Hence a manual implementation...
+        <div
+          className={styles.verticalMarquee}
+          aria-label='Scrolling images of DevFest 2024 Speakers'
+        >
           <ul className={`${styles.marqueeContent} ${styles.marqueeContent_up}`}>
             {[...speakersSlideData.slice(0, 3), ...speakersSlideData.slice(0, 3)].map(
               (speaker, index) => {
@@ -82,18 +87,16 @@ const Hero = () => {
           </ul>
         </div>
       ) : (
-        <div className={styles.horizontalMarquee}>
-          <ul className={styles.marqueeContent}>
-            {[...speakersSlideData, ...speakersSlideData].map((speaker, index) => {
-              return (
-                <li className={styles.marqueeItem} key={index}>
-                  <Image src={speaker.image} fill alt={speaker.name + "'s image"} />
-                  <div className={styles.speakerName}>{speaker.name}</div>
-                </li>
-              );
-            })}
-          </ul>
-        </div>
+        <Marquee speed={20} pauseOnHover aria-label='Scrolling images of DevFest 2024 Speakers'>
+          {speakersSlideData.map((speaker, index) => {
+            return (
+              <div className={`${styles.marqueeItem} ${styles.horizontalMarqueeItem}`} key={index}>
+                <Image src={speaker.image} fill alt={speaker.name + "'s image"} />
+                <div className={styles.speakerName}>{speaker.name}</div>
+              </div>
+            );
+          })}
+        </Marquee>
       )}
     </section>
   );
