+++ src/app/speakers/styles/DaysAndFilter.module.scss
@@ -5,10 +5,10 @@
 .DayFilterWrapper {
   background: $yellow-primary-100;
   padding: 1.5rem 19rem;
+  padding-bottom: 0;
   display: flex;
   flex-direction: column;
   align-items: center;
-  gap: 1.5rem;
 
   button {
     background: inherit;
@@ -23,6 +23,7 @@
     gap: 1.5rem;
     align-items: center;
     justify-content: center;
+    padding: 1.5rem;
 
     .dayBtn {
       @include body2-medium();
@@ -34,18 +35,33 @@
         @include body3-medium();
         padding: 1rem;
       }
+
+      @include media('<=400px') {
+        @include body4-medium();
+        padding: 0.8rem;
+      }
     }
     .activeDayBtn {
       background-color: $core-yellow;
     }
 
-    @include media('<=tablet') {
+    @include media('<=desktop') {
       padding: 0 1.6rem;
+      gap: 0.8rem;
+    }
+
+    @include media('<=tablet') {
       gap: 0.5rem;
+    }
+
+    @include media('<=phone') {
+      padding: 0 1rem;
     }
   }
 
   .FilterBtnsContainer {
+    padding: 3rem 0;
+
     .FilterBtnsFlex {
       display: flex;
       row-gap: 1rem;
@@ -84,12 +100,13 @@
       overflow-x: scroll;
       overscroll-behavior-inline: contain;
       -webkit-overflow-scrolling: touch;
-      padding: 0 2rem;
+      padding: 1.5rem 2rem;
     }
   }
 
   @include media('<=tablet') {
     padding: 1rem 0;
-    gap: 1.5rem;
+    padding-top: 4rem;
+    gap: 0rem;
   }
 }
