+++ src/styles/components/_faqs.scss
@@ -81,7 +81,7 @@
       cursor: pointer;
       justify-content: space-between;
       display: flex;
-      gap: 10px;
+      // gap: 10px;
       flex-direction: column;
       width: 100%;
       max-height: 60px;
@@ -109,8 +109,8 @@
         width: 60%;
         max-width: 517px;
         padding-left: 1rem;
-        opacity: o;
-        transition: opacity ease 2s;
+        opacity: 0;
+        height: 0;
 
         @media (max-width: 768px) {
           width: 80%;
@@ -138,10 +138,13 @@
     }
     .is-opened {
       background: $yellow-primary-90;
+      gap: 10px;
       max-height: 120px;
 
       .item-answer {
         opacity: 1;
+        height: fit-content;
+        transition: opacity ease 100ms;
       }
 
       .arrow-icon {
